requirements.txt
pytest
pytest-dotenv
pandas
sqlalchemy
pydantic
pg8000
psycopg2-binary
opentelemetry-exporter-otlp
opentelemetry-instrumentation-dbapi
opentelemetry-instrumentation-logging
